Remove async functions to improve code testability
