+++ satellite/weather/__init__.py
@@ -3,7 +3,8 @@
 # TODO: docstrings
 import xarray as xr
 
-from .xr_extensions import *
+from .copebr import * # noqa
+from .dsei import * # noqa
 
 
 def load_dataset(file_path: str) -> xr.Dataset:
