+++ satellite/weather/brazil/DSEI/areas.py
