+++ satellite/weather/brazil/DSEI/areas_dsei.cst
