+++ satellite/weather/brazil/__init__.py
