+++ satellite/weather/brazil/municipios.json
@@ -1235,7 +1235,7 @@
         "geocodigo": 3153509,
         "municipio": "Alto Jequitibá",
         "latitude": -20.4208,
-        "longitude": -41.967,
+        "longitude": -41.9670,
         "codigo_uf": 31,
         "fuso_horario": "America\/Sao_Paulo"
     },
@@ -5539,7 +5539,7 @@
         "geocodigo": 1703305,
         "municipio": "Bom Jesus do Tocantins",
         "latitude": -8.96306,
-        "longitude": -48.165,
+        "longitude": -48.1650,
         "codigo_uf": 17,
         "fuso_horario": "America\/Sao_Paulo"
     },
